requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
scikit-learn
joblib
httpx
streamlit